requirements.txt
flask==2.3.2
flask-cors==4.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
twilio==8.10.0
requests==2.31.0
pydantic==2.5.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
openpyxl==3.1.2
pandas==2.0.3
numpy==1.24.3
python-dateutil==2.8.2
pyyaml==6.0
pytest==7.4.0
pytest-cov==4.1.0
sarvamai>=0.1.0
